casts to keep compiler happy
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@32733 72102866-910b-0410-8b05-ffd578937521

+++ Source/GSAttributedString.m
@@ -144,7 +144,7 @@ cacheAttributes(NSDictionary *attrs)
        * in an immutable dictionary that can safely be cached.
        */
       attrs = [[NSDictionary alloc] initWithDictionary: attrs copyItems: NO];
-      GSIMapAddPair(&attrMap, (GSIMapKey)((id)attrs), (GSIMapVal)(unsigned)1);
+      GSIMapAddPair(&attrMap, (GSIMapKey)((id)attrs), (GSIMapVal)(NSUInteger)1);
     }
   else
     {
